Added troubleshooting, added pkgver() in PKGBUILD
There is a new function in the script that should help users to fix common
errors that happen with pacman, it's not complete yet, just a scheme.
Now the PKGBUILD use the pkgver() function to update version number, as the
ArchWiki says it should be done, maybe it's better maybe it's not.