Fixed unusual crash and cache elements counting
Pacaur cache was counted estimating used space, but was excluded from the removing function; now it should be excluded also from the counting.
Removed last '[[ condition ]] && command' structure, a normal 'if .. then' is used instead.
Orphan packages are now saved in array, this way they can be manipulated without restrictions.
Fixed a crash when the user refuse to remove orphan packages, this happened beacuse pacman returns error if a user stops the operation answering 'No' to the confirmation question. Strange behaviour in my opinion, now I have to ignore any error message in that section.
Removed the variable contaning 'sudo' commmand, it was useless with recent versions.
It's not tested yet, but the script contained a variable evaluation that can crash it on 'no Manjaro' systems, now that should be fixed.