Pacman noconfirm and code clean
Now every pacman transaction does not ask for confirmation, this way it is impossible to obtain an error from pacman if everything goes well with the update, the script prrints a list of upgradable packages to let the user choose freely.
Added an 'orphan packages' funtion to clean the code a little bit.
Added 'local' property to some variable that deserve it.
The script in the future will check something before asking to user, for example it already checks if there is any orphan package and prints the list (if there are many), before asking the user if he wants to remove them.
Changed convention for variable names, lower case in most cases.