Added EXIT trap, update to README
An EXIT trap can stop the script before exiting, the user can see error messages
also if cleanupdate is opened from the menu (indipendent terminal window).
Previous behaviour was to quit immediately after the EXIT, and errors were
impossible to read.
The script checks the existence of '/bin/tput' before using it.
The README file now contains a list of dependencies, for users installing it
manually.