Design update, better system update, better logs
Little design update of the main menu and version bumping.
System updates do not depend on checkupdates anymore, only on pacman,
and the script should log any error, the downside is that the script
does not exit on pacman errors, it only reports the system update as "canceled".
Now the log contains the most part of the error messages from different
command: pacman, rm and others.
This last part should be refined a little, for example printing a
separator in logs before any funcion.