Add pipefail option, handle man errors
The script set the 'pipefail' option in bash, this enable error detection in
pipe redirection.
When the manual page is called the system often complains about bad locale
settings, now the script handles this error without exiting, pointing the user
in the right direction with a message and logging the error in the cleanupdate
temporary log.