Fix condition error with pipefail
An 'if' statement has a condition containing a pipe, with the
'pipefail' option this condition is always evaluated to 'false' and the
function does not work, the temporary workaround is to disable the
'pipefail' option for this 'if' statement.
Changed the log size check limit to match the size in the 'journalctl'
command to clean logs.