Remove auto-downgrade option
Remove the check for package newer than repository, I think this is not
a job for a simple maintenance. I added the 'newer than' warning to the list of
troubleshooting options.
I have removed it right now (and not in another commit or branch) because
with 'pkexec' introduction there is a password prompt every time, it becomes
annoying and this call was the less useful.