Remove yaourt, add pamac support
Completely remove support for 'yaourt' pacman helper.
Add initial support for 'pamac', this is not only an AUR helper so it
triggers a complete system upgrade together with the AUR upgrade.
At the moment 'pamac' just replaces 'yaourt' but it probably deserves a
different implementation in order to use it properly both for official
and AUR packages upgrade.