Enable 'pkexec' in TTY too
Even with the problem about 'pkexec' in a TTY
https://gitlab.freedesktop.org/polkit/polkit/issues/17
is still possible to run in, first we start a pkttyagent on the shell
PID and then we kill it before exiting.
At the moment 'journalctl' complains about this, probably because it has
its own polkit agent, this needs to be addressed.